Creating a user should now correctly create a salt And use it to hash the password

+++ src/nl/itopia/corendon/data/Employee.java
@@ -15,8 +15,8 @@ public class Employee {
     public String notes;
     public Role role;
     public Airport airport;
-    public int createDate;
-    public int lastOnline;
+    public long createDate;
+    public long lastOnline;
 
     public Employee(int id) {
         this.id = id;
